app/backend/requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-cors-headers==4.3.1